creates modules
